webpack configured, code divided into modules
