+++ public/index.html
@@ -45,6 +45,6 @@
           </section>
       </div>
     </main>
-    <script src="../src/main.js"></script>
+    <script src="./bundle.js"></script>
   </body>
 </html>
