+++ src/components/content.js
@@ -1,0 +1,68 @@
+// import {getTripEventsItemLayout}
+
+import {getTripEventsItemLayout} from './events';
+
+const getTripInfoLayout = () => `
+<div class="trip-info__main">
+  <h1 class="trip-info__title">Amsterdam &mdash; ... &mdash; Amsterdam</h1>
+  <p class="trip-info__dates">Mar 18&nbsp;&mdash;&nbsp;21</p>
+</div>
+`;
+
+const getTripEventsSortLayout = () => `
+<form class="trip-events__trip-sort  trip-sort" action="#" method="get">
+  <span class="trip-sort__item  trip-sort__item--day">Day</span>
+
+  <div class="trip-sort__item  trip-sort__item--event">
+    <input id="sort-event" class="trip-sort__input  visually-hidden" type="radio" name="trip-sort" value="sort-event" checked>
+    <label class="trip-sort__btn" for="sort-event">Event</label>
+  </div>
+
+  <div class="trip-sort__item  trip-sort__item--time">
+    <input id="sort-time" class="trip-sort__input  visually-hidden" type="radio" name="trip-sort" value="sort-time">
+    <label class="trip-sort__btn" for="sort-time">
+      Time
+      <svg class="trip-sort__direction-icon" width="8" height="10" viewBox="0 0 8 10">
+        <path d="M2.888 4.852V9.694H5.588V4.852L7.91 5.068L4.238 0.00999987L0.548 5.068L2.888 4.852Z"/>
+      </svg>
+    </label>
+  </div>
+
+  <div class="trip-sort__item  trip-sort__item--price">
+    <input id="sort-price" class="trip-sort__input  visually-hidden" type="radio" name="trip-sort" value="sort-price">
+    <label class="trip-sort__btn" for="sort-price">
+      Price
+      <svg class="trip-sort__direction-icon" width="8" height="10" viewBox="0 0 8 10">
+        <path d="M2.888 4.852V9.694H5.588V4.852L7.91 5.068L4.238 0.00999987L0.548 5.068L2.888 4.852Z"/>
+      </svg>
+    </label>
+  </div>
+
+  <span class="trip-sort__item  trip-sort__item--offers">Offers</span>
+</form>
+`;
+
+const getTripDaysContainerLayout = () => `
+<ul class="trip-days">
+  <li class="trip-days__item  day">
+    <div class="day__info">
+      <span class="day__counter">1</span>
+      <time class="day__date" datetime="2019-03-18">MAR 18</time>
+    </div>
+
+    <ul class="trip-events__list">
+      <li class="trip-events__item">
+        ${getTripEventsItemLayout()}
+      </li>
+      <li class="trip-events__item">
+        ${getTripEventsItemLayout()}
+      </li>
+      <li class="trip-events__item">
+        ${getTripEventsItemLayout()}
+      </li>
+    </ul>
+  </li>
+</ul>
+`;
+
+export {getTripInfoLayout, getTripEventsSortLayout, getTripDaysContainerLayout};
