+++ src/components/utils.js
@@ -1,0 +1,5 @@
+const renderElement = (container, layout, position = `beforeend`) => {
+  container.insertAdjacentHTML(position, layout);
+};
+
+export {renderElement};
